movie-service/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.20.1
exceptiongroup==1.1.2
fastapi==0.101.0
databases[postgresql]==0.2.6
pydantic==2.1.1
pydantic-settings==2.0.3
pydantic_core==2.4.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.11.2
httpx==0.11.1
